move submission progress bar listener to own file
