+++ app/src/org/commcare/activities/SyncCapableCommCareActivity.java
@@ -13,6 +13,7 @@ import org.commcare.logging.analytics.GoogleAnalyticsFields;
 import org.commcare.logging.analytics.GoogleAnalyticsUtils;
 import org.commcare.tasks.DataPullTask;
 import org.commcare.tasks.DataSubmissionListener;
+import org.commcare.tasks.FormSubmissionProgressBarListener;
 import org.commcare.tasks.ProcessAndSendTask;
 import org.commcare.tasks.PullTaskResultReceiver;
 import org.commcare.tasks.ResultAndError;
@@ -220,8 +221,8 @@ public abstract class SyncCapableCommCareActivity<T> extends SessionAwareCommCar
 
     public abstract boolean shouldShowSyncItemInActionBar();
 
-    public DataSubmissionListener getListenerForSubmissionProgressBar() {
-        return null;
+    public boolean usesSubmissionProgressBar() {
+        return false;
     }
 
 }
